Fix: link the top-left logo to `logo_home_link`
This fix works by adding missing two missing `.extra`'s when accessing the `logo_home_link` in the `header` template block.

The bug is caused by the first missing `.extra`, which makes Tera assume that `logo_home_link` variable doesn't exist, so the `if` trigger for using it never fires. This means that the logo will always link to the default value of `base_url`.

The second missing `.extra` comes after the `if` fires. It's used to set the value of `logo_link` to `logo_home_link`.