feat: Change RSS icon, do not prepend {{name}}
There's some css rules for which prepending the name of the menu item as
set in the config causes the <a> element to appear shifted down compared
to the other <a> elements of the menu. I might have a look later, but
for now this enables only the appeareance of the RSS icon, discarding
entirely the name of the link.